move files from `views` to `components`
the reason to do this is that the `views` folder will contain the views for the spa
