+++ src/app/components/renderProductDetails.ts
